Move the askChatGPT function
